requirements.txt
flask
requests
yfinance
beautifulsoup4
pandas
matplotlib
plotly
openai
python-dotenv
ibm-watsonx
alpha-vantage